+++ Simulator/gazebo_simulator/CMakeLists.txt
@@ -4,7 +4,7 @@ project(prometheus_gazebo)
 # Load catkin and all dependencies required for this package
 find_package(catkin REQUIRED COMPONENTS
     roscpp
-	rospy
+    rospy
     gazebo_ros
     image_transport
     camera_info_manager
@@ -15,6 +15,9 @@ find_package(catkin REQUIRED COMPONENTS
     gazebo_msgs
     prometheus_msgs
     message_generation
+    geometry_msgs
+    urdf
+    xacro
 )
 find_package(Boost REQUIRED COMPONENTS thread)
 find_package(gazebo REQUIRED)
@@ -33,9 +36,11 @@ generate_messages(DEPENDENCIES
 catkin_package(
     CATKIN_DEPENDS
         roscpp
+        rospy
         gazebo_ros
 	std_msgs
 	message_runtime
+        geometry_msgs
 )
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${GAZEBO_CXX_FLAGS}")
@@ -72,16 +77,18 @@ target_link_libraries(qrcode_control ${catkin_LIBRARIES})
 #target_link_libraries(PriusHybridPlugin ${GAZEBO_LIBRARIES} ${IGNITION-MSGS_LIBRARIES})
 #add_dependencies(PriusHybridPlugin ${catkin_EXPORTED_TARGETS})
 
-catkin_install_python(PROGRAMS scripts/joystick_translator
+catkin_install_python(PROGRAMS nodes/joystick_translator
                       DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION})
 
   install(PROGRAMS 
-   scripts/keyboard_control_px4.py 
+   nodes/keyboard_control_px4.py 
    DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
   install(PROGRAMS 
-   scripts/qrcode_control.py 
+   nodes/qrcode_control.py 
+   nodes/turtlebot3_teleop_key
    DESTINATION ${CATKIN_PACKAGE_BIN_DESTINATION}
 )
 install(DIRECTORY models DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
 install(DIRECTORY worlds DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
+install(DIRECTORY  models/turtlebot3_description/meshes models/turtlebot3_description/rviz models/turtlebot3_description/urdf DESTINATION ${CATKIN_PACKAGE_SHARE_DESTINATION})
