+++ Simulator/gazebo_simulator/launch/iris_with_realsense.launch
@@ -28,7 +28,7 @@
         <include file="$(find mavros)/launch/node.launch">
             <arg name="pluginlists_yaml" value="$(find mavros)/launch/px4_pluginlists.yaml" />
             <!-- Need to change the config file to get the tf topic and get local position in terms of local origin -->
-            <arg name="config_yaml" value="$(find simulation)/resource/px4_config.yaml" />
+            <arg name="config_yaml" value="$(find prometheus_gazebo)/config/px4_config.yaml" />
             <arg name="fcu_url" value="$(arg fcu_url)" />
             <arg name="gcs_url" value="$(arg gcs_url)" />
             <arg name="tgt_system" value="$(arg tgt_system)" />
@@ -39,7 +39,7 @@
     <!-- Launch Gazebo -->
     <include file="$(find gazebo_ros)/launch/empty_world.launch">
       <arg name="gui" value="$(arg gui)"/>
-      <arg name="world_name" value="$(find iris_gazebo)/worlds/cloister.world"/>
+      <arg name="world_name" value="$(find prometheus_gazebo)/worlds/cloister.world"/>
     </include>
 
     <!-- Spawn vehicle model -->
@@ -47,6 +47,6 @@
           args="-sdf -database iris_realsense_camera -model $(arg vehicle)">
     </node>
     
-    <node pkg="iris_gazebo" type="keyboard_control_px4.py" name="keyboard_control_px4" output="screen" launch-prefix="gnome-terminal --tab -e">
+    <node pkg="prometheus_gazebo" type="keyboard_control_px4.py" name="keyboard_control_px4" output="screen" launch-prefix="gnome-terminal --tab -e">
     </node>
 </launch>
