+++ Simulator/gazebo_simulator/launch/landing_with_qrcode.launch
@@ -7,6 +7,10 @@
     <arg name="tgt_system" default="1" />
     <arg name="tgt_component" default="1" />
     <arg name="vehicle" default="iris"/>
+    <arg name="model_tut" default="$(env TURTLEBOT3_MODEL)" doc="model type [burger, waffle, waffle_pi]"/>
+    <arg name="x_pos" default="5.0"/>
+    <arg name="y_pos" default="0.0"/>
+    <arg name="z_pos" default="0.0"/>
 
     <param name="use_sim_time" value="true" />
 
@@ -37,19 +41,22 @@
     </group>
 
     <!-- Launch Gazebo -->
-    <include file="$(find gazebo_ros)/launch/empty_world.launch">
-      <arg name="gui" value="$(arg gui)"/>
-      <arg name="world_name" value="$(find prometheus_gazebo)/worlds/qrcode.world"/>
-    </include>
-
-    <!-- Move the QRCode -->
-    <node pkg="prometheus_gazebo" type="qrcode_control.py" name="qrcode_control">
-    </node>
+  <include file="$(find gazebo_ros)/launch/empty_world.launch">
+    <arg name="world_name" value="$(find prometheus_gazebo)/worlds/qrcode.world"/>
+    <arg name="paused" value="false"/>
+    <arg name="use_sim_time" value="true"/>
+    <arg name="gui" value="true"/>
+    <arg name="headless" value="false"/>
+    <arg name="debug" value="false"/>
+  </include>
+   <param name="robot_description" command="$(find xacro)/xacro --inorder $(find prometheus_gazebo)/models/turtlebot3_description/urdf/turtlebot3_$(arg model_tut).urdf.xacro" />
 
     <!-- Spawn vehicle model -->
     <node name="spawn_model" pkg="gazebo_ros" type="spawn_model" output="screen"
           args="-sdf -database iris_realsense_camera_qrcode -model $(arg vehicle)">
     </node>
+
+    <node pkg="gazebo_ros" type="spawn_model" name="spawn_urdf" args="-urdf -model turtlebot3_$(arg model) -x $(arg x_pos) -y $(arg y_pos) -z $(arg z_pos) -param robot_description" />
     
     <node pkg="prometheus_gazebo" type="keyboard_control_px4.py" name="keyboard_control_px4" output="screen" launch-prefix="gnome-terminal --tab -e">
     </node>
