+++ Simulator/gazebo_simulator/media/materials/scripts/qrcode_cube.material
@@ -1,0 +1,15 @@
+material Cube/QRCode
+{
+  technique
+  {
+    pass
+    {
+      texture_unit
+      {
+        texture qrcode2.PNG
+        scale 1 1
+      }
+    }
+  }
+}
+
