+++ Simulator/gazebo_simulator/models/qrcode_cube/qrcode_cube.sdf
@@ -2,7 +2,7 @@
 <sdf version="1.5">
   <model name="qrcode_cube">
     <link name="link">
-      <pose>1.0 0.0 0.05 0 0 0</pose>
+      <pose>5.0 0.0 0.05 0 0 0</pose>
       <inertial>
         <mass>0.5167</mass>
         <inertia>
@@ -18,7 +18,7 @@
       <collision name="collision">
         <geometry>
           <box>
-            <size>0.6 0.6 0.04</size>
+            <size>2 2 2</size>
           </box>
         </geometry>
         <surface>
@@ -33,11 +33,10 @@
             </ode>
           </contact>
           <friction>
-            <torsional>
-              <coefficient>1.0</coefficient>
-              <use_patch_radius>0</use_patch_radius>
-              <surface_radius>0.01</surface_radius>
-            </torsional>
+	    <ode>
+ 	      <mu>0.2</mu>
+	      <mu2>0.2</mu2>
+            </ode>
           </friction>
         </surface>
       </collision>
@@ -45,7 +44,7 @@
       <visual name="visual">
         <geometry>
           <box>
-            <size>0.6 0.6 0.01</size>
+            <size>2 2 2</size>
           </box>
         </geometry>
         <material>
