+++ Simulator/gazebo_simulator/worlds/qrcode.world
@@ -1,5 +1,5 @@
 <?xml version="1.0" ?>
-<sdf version="1.5">
+<sdf version="1.4">
   <world name="default">
     <!-- A global light source -->
     <include>
@@ -9,32 +9,24 @@
     <include>
       <uri>model://ground_plane</uri>
     </include>
-    <include>
-      <uri>model://lawn_plane</uri>
-    </include>
-    <include>
-      <uri>model://qrcode_cube</uri>
-    </include>
-    <physics name='default_physics' default='0' type='ode'>
-      <gravity>0 0 -9.8066</gravity>
+    <physics type='ode'>
       <ode>
         <solver>
           <type>quick</type>
-          <iters>10</iters>
+          <iters>150</iters>
           <sor>1.3</sor>
           <use_dynamic_moi_rescaling>0</use_dynamic_moi_rescaling>
         </solver>
         <constraints>
-          <cfm>0</cfm>
+          <cfm>0.00001</cfm>
           <erp>0.2</erp>
-          <contact_max_correcting_vel>100</contact_max_correcting_vel>
-          <contact_surface_layer>0.001</contact_surface_layer>
+          <contact_max_correcting_vel>1000</contact_max_correcting_vel>
+          <contact_surface_layer>0.01</contact_surface_layer>
         </constraints>
       </ode>
       <max_step_size>0.004</max_step_size>
-      <real_time_factor>1</real_time_factor>
+      <real_time_factor>1000.0</real_time_factor>
       <real_time_update_rate>250</real_time_update_rate>
-      <magnetic_field>6.0e-6 2.3e-5 -4.2e-5</magnetic_field>
     </physics>
   </world>
 </sdf>
