+++ 日志.md
@@ -1,0 +1,9 @@
+###### 2020.1.30
+
+参考项目：https://github.com/Lee0326/sim2real_drone_racing 将其仿真环境移植;
+
+加入穿越机仿真的模型文件到仿真器存放路径，其中主要是赛道gate的三位模型文件和纹理文件;
+
+加入穿越机赛道文件，可以通过gazebo打开，验证ok;
+
+![image-20200130114224391](日志.assets/image-20200130114224391.png)
